% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clogit_perm.R
\name{clogit_perm}
\alias{clogit_perm}
\title{Conditional logit with permutation}
\usage{
clogit_perm(
  nperm = 1000,
  formula,
  data,
  ...,
  ncpus = parallel::detectCores(),
  parallel_args = list()
)
}
\arguments{
\item{nperm}{Integer. Number of permutations}

\item{formula, data, ...}{Parameters passed to \code{\link[survival:clogit]{survival::clogit()}}}

\item{ncpus}{Integer. Number or cores.}

\item{parallel_args}{Parameters passed to \code{\link[parallel:makeCluster]{parallel::makeCluster()}}.}
}
\description{
Conditional logit with permutation
}
