% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_candidates}
\alias{find_candidates}
\title{Find permutation candidates for each row}
\usage{
find_candidates(dates, unit_id, incident_id, ematch, window)
}
\arguments{
\item{dates}{Vector of dates}

\item{unit_id}{Integer vector of individuals' ids (to avoid matching with
self).}

\item{incident_id}{Integer vector of the corresponding incident (to avoid
matching to the same incident).}

\item{ematch}{Integer matrix of variables over which exact match is desired.}

\item{window}{Integer scalar. Number of units of time to consider when
identifying candidates (see details)}
}
\value{
A list of integer vectors (starting from zero) indicating the position of
the potential permutation.
}
\description{
Find permutation candidates for each row
}
\details{
For each row \code{i} in the data, a row \code{j} is a possible permutation candidate if:
\itemize{
\item \code{abs(date[i] - date[j]) <= window},
\item \code{unit_id} are different,
\item \code{incident_id} are different, and
\item \code{all(ematch[i, ] == ematch[j,])}
}
}
