% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_events}
\alias{sim_events}
\alias{sim_events2}
\title{Simulate Police Force Events}
\usage{
sim_events(
  nevents,
  nofficers,
  min_per_event = 1,
  max_per_event = 5,
  min_year = 0,
  max_year = 10,
  min_rate = 5,
  max_rate = 5,
  female_par = -0.5,
  years_par = -0.5,
  rho_par = 0,
  exposure_par = 0.5,
  nsims = 1,
  seed = sample.int(.Machine$integer.max, 1)
)

sim_events2(
  incidentid,
  officerid,
  female,
  years,
  female_par = -0.5,
  years_par = . - 5,
  rho_par = 0.5,
  exposure_par = 0.5,
  nsims = 1,
  seed = sample.int(.Machine$integer.max, 1)
)
}
\arguments{
\item{nevents, nofficers}{Integers. Number of events and officers to simulate.}

\item{min_per_event, max_per_event}{Integers. Lower and upper bounds for the
number of officers in the event.}

\item{min_year, max_years}{Integers. Lower and upper bounds for the number
of years of experience of the officers.}

\item{min_rate, max_rate}{Doubles. Lower and upper bounds for the reaction
rates (see details).}

\item{female_par, years_par, rho_par, exposure_par}{Doubles. Parameters (coefficients) for
the logistic probabilities.}

\item{seed}{Integer. Seed for the pseudo-number generation.}

\item{incidentid, officerid}{Integer vectors. Values for the incident and
officer id.}

\item{female, years}{Logical and integer vectors, respectively. Features
of the officers.}
}
\value{
A data frame with the following columns
\itemize{
\item Officer id
\item Whether the officer is female
\item Years of experience
\item Incident id
\item Whether the officer pointed a gun
}

Each row represents one report per officer involved in the event.
}
\description{
This function generates data similar to that featured in the paper. Events
are drawn at random, as the number of officers per event. The outcome variable,
whether the officer points his gun or not, is drawn sequentially as a poisson
process.
}
\details{
The simulation process goes as follow:
\enumerate{
\item The officers are simulated. Female ~ Bernoulli(0.5),
Action rate ~ Unif(min_rate, max_rate),
Years of experience ~ Discrete Unif\link{min_years, max_year}
\item Events are simulated, each event has a nofficers ~ Discrete Unif\link{min_per_event, max_per_event}
Once the event is done, a sequence of reaction is given by each officers'
action rate (Poisson process). Whether an officer points or not is set by
a logistic model

point ~ female + years of experience + has any pointed? + previous exposure

The corresponding parameters are as specified by the user. Events are simulated
one at a time.
}

In the case of \code{sim_events2}, the user can pass predefined events and
officers and use those to simulate each officers' reactions.
}
\examples{
x <- simulate_njforce(1000, 400)
}
